+++ apps/docs/app/routes/_base/route.tsx
@@ -27,6 +27,7 @@ export default function BaseLayout() {
         marginX={[3, null, 6, 10]}
         marginBottom={["60px", null, "120px", "180px"]}
         maxWidth="924px"
+        justifyContent={"center"}
         ref={contentRef}
       >
         <Outlet />