+++ apps/docs/app/routes/playground/route.tsx
@@ -1,4 +1,4 @@
-import { Stack } from "@vygruppen/spor-react";
+import { Stack, PressableCard } from "@vygruppen/spor-react";
 import { useEffect, useState } from "react";
 import { LivePreview } from "react-live";
 import { LiveEditor } from "~/features/portable-text/interactive-code/LiveEditor";
@@ -25,6 +25,7 @@ export default function PlaygroundPage() {
     setPlaygroundData(newCode);
     localStorage.setItem("playgroundData", newCode);
   };
+
   return (
     <LiveProvider code={playgroundData}>
       <Stack spacing={2} id="content">