+++ packages/spor-react/src/card/index.tsx
@@ -1,3 +1,1 @@
 export * from "./Card";
-export * from "./StaticCard";
-export * from "./PressableCard";