+++ packages/spor-react/src/nudge/Nudge.tsx
@@ -74,7 +74,7 @@ const EXPIRATION_DELAY = 1000 * 60 * 60 * 24 * 30; // 30 days
  *  name="my-nudge"
  *  content="Check out this enormous new feature!"
  * >
- *   <Card variant="base" padding={2} width="fit-content">My new feature</Card>
+ *   <StaticCard variant="base" padding={2} width="fit-content">My new feature</StaticCard>
  * </Nudge>
  * ```
  */