+++ packages/spor-react/src/theme/components/radio-card.ts
@@ -2,46 +2,41 @@ import { defineStyleConfig } from "@chakra-ui/react";
 import { baseBackground, baseBorder, baseText } from "../utils/base-utils";
 import { floatingBackground, floatingBorder } from "../utils/floating-utils";
 import { focusVisibleStyles } from "../utils/focus-utils";
-import { accentBackground } from "../utils/accent-utils";
 
 const config = defineStyleConfig({
-  baseStyle: (props) => ({
+  baseStyle: (props: any) => ({
     appearance: "none",
     border: "none",
     overflow: "hidden",
     fontSize: "inherit",
     display: "block",
-    borderRadius: "md",
-    cursor: "pointer",
+    borderRadius: "sm",
     ...focusVisibleStyles(props),
+    _checked: {
+      outline: "1px solid",
+      outlineColor: "greenHaze",
+      ...floatingBackground("active", props),
+      _hover: {
+        ...floatingBackground("active", props),
+      },
+    },
     _disabled: {
-      ...baseBackground("disabled", props),
-      ...baseText("disabled", props),
-      outline: "none",
       pointerEvents: "none",
+      ...baseBackground("disabled", props),
+      ...baseBorder("disabled", props),
+      ...baseText("disabled", props),
     },
   }),
   variants: {
     base: (props) => ({
+      ...baseBackground("default", props),
       ...baseBorder("default", props),
       _hover: {
+        ...baseBackground("hover", props),
         ...baseBorder("hover", props),
       },
       _active: {
         ...baseBackground("active", props),
-        ...baseBorder("active", props),
-      },
-    }),
-    accent: (props) => ({
-      ...accentBackground("default", props),
-      boxShadow: "sm",
-      _hover: {
-        ...accentBackground("hover", props),
-        boxShadow: "sm",
-      },
-      _active: {
-        ...accentBackground("active", props),
-        boxShadow: "none",
       },
     }),
     floating: (props) => ({
@@ -54,7 +49,13 @@ const config = defineStyleConfig({
       },
       _active: {
         ...floatingBackground("active", props),
-        boxShadow: "none",
+        ...floatingBorder("active", props),
+      },
+      _checked: {
+        _hover: {
+          outline: "1px solid",
+          outlineColor: "silver",
+        },
       },
     }),
   },